requirements.txt
pycryptodome==3.14.1
PyJWT==2.3.0
PySimpleGUI==4.59.0
